Handle Streams in json reader
